Update to work with Materials
Changed constructor so that a list of Material objects is stored for each Multilayer_Sail object instead of merely a string of material names. Using this, the _find_structure() method also had to be edited to work with the newly introduced Material objects. Edited sections marked out; enclosed by ###. Reasonably large non-trivial edits, so creating a new branch for the edit in case things break