Included code to determine eq temps
Rewrote _find_absorptance() to determine the absorptance of a multilayer_sail depending on the abs_coeffs of the materials that compose it. Added functionality to include a wavelength to allow for finding absorptances at different wavelengths (not just the initial laser wavelength, as previously written)

Wrote _find_eq_temps_given_abs_coeff() which allows for the finding of sail equilibrium temperatures if the absorption coefficients of the materials that compose it are provided. Updated from previous version by using a Newton method to implicitly solve for equilibrium temperature instead of poorly time-optimised bisection method. Large changes made to work with material Objects.

Note: future use of find_eq_temps_given_abs_coeff() should be through a file which determines these equilibrium temperatures over a range of abs_coeffs for a single material. Current implementation is used for generality and allows for user manipulation for additional use where that may exist